+++ lib/symmetric_encryption/keystore.rb
@@ -79,8 +79,9 @@ module SymmetricEncryption
           version:     version,
           environment: environment
         }
-        args[:key_path] = ::File.dirname(config[:key_filename]) if config.key?(:key_filename)
-        new_data_key    = keystore_class.generate_data_key(**args)
+        args[:key_path]    = ::File.dirname(config[:key_filename]) if config.key?(:key_filename)
+        args[:permissions] = config[:permissions] if config.key?(:permissions)
+        new_data_key       = keystore_class.generate_data_key(**args)
 
         # Add as second key so that key can be published now and only used in a later deploy.
         if rolling_deploy
@@ -125,7 +126,8 @@ module SymmetricEncryption
           environment: environment,
           dek:         key
         }
-        args[:key_path] = ::File.dirname(config[:key_filename]) if config.key?(:key_filename)
+        args[:key_path]    = ::File.dirname(config[:key_filename]) if config.key?(:key_filename)
+        args[:permissions] = config[:permissions] if config.key?(:permissions)
 
         new_config                     = keystore_class.generate_data_key(**args)
         # Only carry over these settings when they were set explicitly, otherwise the